--- src/hotspot/share/gc/serial/serialFullGC.inline.hpp
@@ -1,65 +1,0 @@
-/*
- * Copyright (c) 2000, 2024, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- *
- */
-
-#ifndef SHARE_GC_SERIAL_SERIALFULLGC_INLINE_HPP
-#define SHARE_GC_SERIAL_SERIALFULLGC_INLINE_HPP
-
-#include "gc/serial/serialFullGC.hpp"
-
-#include "classfile/classLoaderData.inline.hpp"
-#include "classfile/javaClasses.inline.hpp"
-#include "gc/shared/continuationGCSupport.inline.hpp"
-#include "gc/serial/serialStringDedup.hpp"
-#include "memory/universe.hpp"
-#include "oops/markWord.hpp"
-#include "oops/access.inline.hpp"
-#include "oops/compressedOops.inline.hpp"
-#include "oops/oop.inline.hpp"
-#include "utilities/align.hpp"
-#include "utilities/stack.inline.hpp"
-
-template <class T> inline void SerialFullGC::adjust_pointer(T* p) {
-  T heap_oop = RawAccess<>::oop_load(p);
-  if (!CompressedOops::is_null(heap_oop)) {
-    oop obj = CompressedOops::decode_not_null(heap_oop);
-    assert(Universe::heap()->is_in(obj), "should be in heap");
-
-    if (obj->is_forwarded()) {
-      oop new_obj = obj->forwardee();
-      assert(is_object_aligned(new_obj), "oop must be aligned");
-      RawAccess<IS_NOT_NULL>::oop_store(p, new_obj);
-    }
-  }
-}
-
-template <typename T>
-void AdjustPointerClosure::do_oop_work(T* p)           { SerialFullGC::adjust_pointer(p); }
-inline void AdjustPointerClosure::do_oop(oop* p)       { do_oop_work(p); }
-inline void AdjustPointerClosure::do_oop(narrowOop* p) { do_oop_work(p); }
-
-inline size_t SerialFullGC::adjust_pointers(oop obj) {
-  return obj->oop_iterate_size(&SerialFullGC::adjust_pointer_closure);
-}
-
-#endif // SHARE_GC_SERIAL_SERIALFULLGC_INLINE_HPP
